changed files according to description given by prof
