+++ code/2/exercise2.py
@@ -1,0 +1,15 @@
+""" Move the function call back to the bottom and move the definition of
+print_lyrics after the definition of repeat_lyrics. What happens when you run
+this program? """
+
+
+def exercise2():
+    def repeat_lyrics():
+        print_lyrics()
+        print_lyrics()
+
+    def print_lyrics():
+        print "I'm a lumberjack, and I'm okay."
+        print "I sleep all night and I work all day."
+
+    repeat_lyrics()
