+++ code/2/exercise3.py
@@ -3,7 +3,8 @@ parameter and prints the string with enough leading spaces so that the last
 letter of the string is in column 70 of the display. """
 
 
-def right_justify(s):
-    print "%s%s" % (70*" ", s)
+def exercise3():
+    def right_justify(s):
+        print "%s%s" % (70*" ", s)
 
-right_justify("raghav")
+    right_justify("raghav")
