+++ code/2/think3.py
@@ -1,0 +1,11 @@
+import exercise1 as e1, exercise2 as e2, exercise3 as e3, exercise4 as e4, \
+    exercise5 as e5
+
+try:
+    e1.exercise1()
+except Exception as e:
+    print e
+e2.exercise2()
+e3.exercise3()
+e4.exercise4()
+e5.exercise5()
